+++ lib/src/helper/entity.dart
@@ -1,0 +1,9 @@
+import 'package:flutter_smart_dialog/src/strategy/action.dart';
+
+class DialogInfo {
+  DialogInfo(this.action, this.backDismiss);
+
+  final DialogAction action;
+
+  final bool backDismiss;
+}