+++ lib/src/smart_dialog.dart
@@ -4,25 +4,20 @@ import 'package:flutter/material.dart';
 import 'package:flutter_smart_dialog/src/widget/loading_widget.dart';
 
 import 'helper/config.dart';
-import 'helper/proxy.dart';
+import 'helper/dialog_proxy.dart';
 
 class SmartDialog {
-  ///SmartDialog相关配置,使用Config管理
-  Config config = DialogProxy.instance.config;
-
-  ///-------------------------私有类型，不对面提供修改----------------------
-  static SmartDialog? _instance;
-
-  factory SmartDialog() => instance;
-
-  static SmartDialog get instance => _instance ??= SmartDialog._internal();
-
-  SmartDialog._internal();
-
-  ///使用自定义布局
+  /// SmartDialog global config
   ///
-  /// 使用'Temp'为后缀的属性，在此处设置，并不会影响全局的属性，未设置‘Temp’为后缀的属性，
-  /// 则默认使用Config设置的全局属性
+  /// SmartDialog全局配置
+  static Config config = DialogProxy.instance.config;
+
+  /// custom dialog
+  /// attributes that use 'Temp' as the suffix will not affect the global config when used here;
+  /// attributes that do not use ‘Temp' as the suffix will default to the global config
+  ///
+  /// 自定义dialog
+  /// 使用'Temp'为后缀的属性，在此处使用，并不会影响全局的属性；未使用‘Temp’为后缀的属性，会默认使用Config设置的全局属性
   static Future<void> show({
     required Widget widget,
     AlignmentGeometry? alignmentTemp,
@@ -33,27 +28,29 @@ class SmartDialog {
     Color? maskColorTemp,
     Widget? maskWidgetTemp,
     bool? clickBgDismissTemp,
-    //overlay弹窗消失回调
     VoidCallback? onDismiss,
     String? tag,
+    bool? backDismiss,
   }) {
     return DialogProxy.instance.show(
       widget: widget,
-      alignment: alignmentTemp ?? instance.config.alignment,
-      isPenetrate: isPenetrateTemp ?? instance.config.isPenetrate,
-      isUseAnimation: isUseAnimationTemp ?? instance.config.isUseAnimation,
-      animationDuration:
-          animationDurationTemp ?? instance.config.animationDuration,
-      isLoading: isLoadingTemp ?? instance.config.isLoading,
-      maskColor: maskColorTemp ?? instance.config.maskColor,
-      maskWidget: maskWidgetTemp ?? instance.config.maskWidget,
-      clickBgDismiss: clickBgDismissTemp ?? instance.config.clickBgDismiss,
+      alignment: alignmentTemp ?? config.alignment,
+      isPenetrate: isPenetrateTemp ?? config.isPenetrate,
+      isUseAnimation: isUseAnimationTemp ?? config.isUseAnimation,
+      animationDuration: animationDurationTemp ?? config.animationDuration,
+      isLoading: isLoadingTemp ?? config.isLoading,
+      maskColor: maskColorTemp ?? config.maskColor,
+      maskWidget: maskWidgetTemp ?? config.maskWidget,
+      clickBgDismiss: clickBgDismissTemp ?? config.clickBgDismiss,
       onDismiss: onDismiss,
       tag: tag,
+      backDismiss: backDismiss ?? true,
     );
   }
 
-  ///提供loading弹窗
+  /// provide loading dialog
+  ///
+  /// 提供loading弹窗
   static Future<void> showLoading({
     String msg = 'loading...',
     Color background = Colors.black,
@@ -65,22 +62,25 @@ class SmartDialog {
     Color? maskColorTemp,
     Widget? maskWidgetTemp,
     Widget? widget,
+    bool? backDismiss,
   }) {
     return DialogProxy.instance.showLoading(
       widget: widget ?? LoadingWidget(msg: msg, background: background),
       clickBgDismiss: clickBgDismissTemp ?? false,
       isLoading: isLoadingTemp ?? true,
-      maskColor: maskColorTemp ?? instance.config.maskColor,
-      maskWidget: maskWidgetTemp ?? instance.config.maskWidget,
-      isPenetrate: isPenetrateTemp ?? instance.config.isPenetrate,
-      isUseAnimation: isUseAnimationTemp ?? instance.config.isUseAnimation,
-      animationDuration:
-          animationDurationTemp ?? instance.config.animationDuration,
+      maskColor: maskColorTemp ?? config.maskColor,
+      maskWidget: maskWidgetTemp ?? config.maskWidget,
+      isPenetrate: isPenetrateTemp ?? config.isPenetrate,
+      isUseAnimation: isUseAnimationTemp ?? config.isUseAnimation,
+      animationDuration: animationDurationTemp ?? config.animationDuration,
+      backDismiss: backDismiss ?? true,
     );
   }
 
-  /// 提供toast示例
+  /// provide toast
+  /// alignment：control the location of toast
   ///
+  /// 提供toast
   /// alignment：控制toast出现的位置
   static Future<void> showToast(
     String msg, {
@@ -100,10 +100,18 @@ class SmartDialog {
   /// 1：only close dialog
   /// 2：only close toast
   /// 3：only close loading
-  /// 4：both close
+  /// other：all close
   ///
   /// tag：the dialog for setting the 'tag' can be closed
-  static Future<void> dismiss({int closeType = 0,String? tag}) async {
+  ///
+  /// 0：关闭dialog或者loading
+  /// 1：仅关闭dialog
+  /// 2：仅关闭toast
+  /// 3：仅关闭loading
+  /// other：全关闭
+  ///
+  /// tag：设置了tag的dialog，会被针对性关闭
+  static Future<void> dismiss({int closeType = 0, String? tag}) async {
     DialogProxy.instance.dismiss(closeType: closeType, tag: tag);
   }
 }