+++ CHANGELOG.md
@@ -1,9 +1,16 @@
+# [3.0.x]
+
+- support close the dialog when you pop route
+- support open multi dialog
+
 # [2.3.x]
+
 - optimize function
 - solve problem of keyboard shelter toast
 - optimize toast. remove isDefaultDismissType api
 
 # [2.1.x]
+
 - reconstruction of the underlying logic
 - adjust toast default duration
 - add maskWidget param
@@ -13,50 +20,62 @@
 - migrate null-safety
 - adapt flutter 2.0
 
-
 # [1.3.x]
+
 - Improve toast display
 - Improve toast customization function
 
 # [1.2.x]
+
 - loadingDialog perfect parameter settings
 - adjust and use show method
 - add dismiss callback
 
 # [1.1.x]
+
 - simplified use
 - fix bug, adjust the default value of clickBgDismiss attribute to true
 
 # [1.0.3] [1.0.5]
+
 - improve the usage details in the document
 
 # [1.0.1]
+
 - perfect some function
 
 # [0.1.7]
+
 - perfect description
 
 # [0.1.6]
+
 - short descriptio
 
 # [0.1.5]
+
 - perfect description
 
 # [0.1.3]
+
 - adjust code of example
 
 # [0.1.2]
+
 - add example
 
 # [0.1.1]
+
 - remove some constructors
 
 # [0.1.0]
+
 - property debugging is complete
 
 # [0.0.4]
+
 - dealing with the problem of package name
 
-
 # [0.0.1]
+
 - init